refactor tablestore file
