+++ python/datashaper/datashaper/table_store/__init__.py
@@ -1,0 +1,21 @@
+#
+# Copyright (c) Microsoft. All rights reserved.
+# Licensed under the MIT license. See LICENSE file in the project.
+#
+"""The tablestore module contains the table store classes used by the datashaper."""
+from .disk_table_store import DiskTableStore
+from .in_memory_table_store import InMemoryTableStore
+from .table_store import TableStore
+from .types import ColumnMetadata, ColumnStats, Table, TableContainer, TableMetadata
+
+
+__all__ = [
+    "TableStore",
+    "TableContainer",
+    "TableMetadata",
+    "ColumnMetadata",
+    "ColumnStats",
+    "Table",
+    "InMemoryTableStore",
+    "DiskTableStore",
+]
